Update code in branch split_func
