+++ src/import_lib.py
@@ -1,0 +1,11 @@
+from selenium import webdriver
+from selenium.webdriver.chrome.options import Options
+from selenium.common.exceptions import NoSuchElementException
+from selenium.webdriver.common.by import By
+from bs4 import BeautifulSoup
+import pandas as pd
+import requests
+import numpy as np
+import lxml.html as lh
+
+print("Libraries imported")
