+++ src/main.py
@@ -1,5 +1,8 @@
-import liquor_search_scraping
+import import_lib
+import liquor_search_list_scraping
+import liquor_search_inv_scraping
 
 print("Running scrape_olcc_whiskey_list & scrape_olcc_whiskey_inv")
-print(liquor_search_scraping.scrape_olcc_whiskey_list("http://www.oregonliquorsearch.com/"))
-print(liquor_search_scraping.scrape_olcc_whiskey_inv("http://www.oregonliquorsearch.com/"))
+import_lib
+print(liquor_search_list_scraping.scrape_olcc_whiskey_list("http://www.oregonliquorsearch.com/"))
+print(liquor_search_inv_scraping.scrape_olcc_whiskey_inv("http://www.oregonliquorsearch.com/"))
